# src/app/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Thesis API",
    "description": "API documentation for Fazrul Anugrah Sahidin's thesis project",
    "version": "0.1.0",
    "contact": {
      "email": "fazrul.anugrah@gmail.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ],
  "paths": {
...
    "/students": {
      "get": {
        "tags": ["Student"],
        "summary": "Get student list with filter",
        "description": "Get all students, optionally filtered by query parameters such as name, email, studentId, studyProgramId, or entryYear",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/KeywordsParam" },
          {
            "name": "studyProgramId",
            "in": "query",
            "description": "Filter by study program ID",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "entryYear",
            "in": "query",
            "description": "Filter by entry year (angkatan)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1900
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of students retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "number" },
                          "name": { "type": "string" },
                          "email": { "type": "string" },
                          "studentId": { "type": "string" },
                          "studyProgramId": { "type": "integer" }
                        }
                      }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Student"],
        "summary": "Create student",
        "description": "Create a new student",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "studentId": { "type": "string" },
                  "studyProgramId": { "type": "integer" }
                },
                "required": ["name", "email", "studentId", "studyProgramId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "studentId": { "type": "string" },
                        "studyProgramId": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/advisors": {
      "get": {
        "tags": ["Advisor"],
        "summary": "Get advisor list",
        "description": "Get all advisors with optional pagination and keyword search",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/KeywordsParam" }
        ],
        "responses": {
          "200": {
            "description": "Advisor list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Advisor" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Advisor"],
        "summary": "Create advisor",
        "description": "Create a new advisor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdvisorCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Advisor created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Items per page",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      },
      "KeywordsParam": {
        "name": "keywords",
        "in": "query",
        "description": "Free-text search keywords",
        "required": false,
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPrevPage": { "type": "boolean", "example": false }
        },
        "required": ["total", "page", "limit", "totalPages", "hasNextPage", "hasPrevPage"]
      },
      "Advisor": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "userId", "createdAt", "updatedAt"]
      },
      "AdvisorCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "type": { "type": "string", "enum": ["HeadOfDepartment", "HeadOfStudyProgram"] },
          "studyProgramId": { "type": "integer", "nullable": true }
        },
        "required": ["name", "email", "type"]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}